+++ packages/react-app/app/layout.tsx
@@ -1,0 +1,18 @@
+import { Providers } from './providers'
+import '../styles/globals.css'
+
+export default function RootLayout({
+  children,
+}: {
+  children: React.ReactNode
+}) {
+  return (
+    <html lang="en">
+      <body>
+        <Providers>
+          {children}
+        </Providers>
+      </body>
+    </html>
+  )
+}
