+++ packages/react-app/components/HomePage.tsx
@@ -1,0 +1,37 @@
+'use client'
+
+import { useEffect, useState } from 'react'
+import { useAccount } from 'wagmi'
+
+export default function HomePage() {
+  const [userAddress, setUserAddress] = useState('')
+  const [isMounted, setIsMounted] = useState(false)
+  const { address, isConnected } = useAccount()
+
+  useEffect(() => {
+    setIsMounted(true)
+  }, [])
+
+  useEffect(() => {
+    if (isConnected && address) {
+      setUserAddress(address)
+    }
+  }, [address, isConnected])
+
+  if (!isMounted) {
+    return null
+  }
+
+  return (
+    <div className="flex flex-col justify-center items-center">
+      <div className="h1">
+        There you go... a canvas for your next Celo project!
+      </div>
+      {isConnected ? (
+        <div className="h2 text-center">Your address: {userAddress}</div>
+      ) : (
+        <div>No Wallet Connected</div>
+      )}
+    </div>
+  )
+}
