+++ packages/react-app/next.config.js
@@ -4,7 +4,11 @@ const nextConfig = {
   swcMinify: true,
   webpack: (config) => {
     config.resolve.fallback = {
-      fs: false
+      ...config.resolve.fallback,
+      fs: false,
+      net: false,
+      tls: false,
+      encoding: false
     }
     return config
   }
