+++ packages/react-app/package.json
@@ -9,29 +9,31 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@headlessui/react": "^1.7.14",
-    "@heroicons/react": "^2.0.18",
-    "@rainbow-me/rainbowkit": "^2.0.2",
-    "@tanstack/react-query": "^5.28.6",
+    "@headlessui/react": "^1.7.19",
+    "@heroicons/react": "^2.1.5",
+    "@rainbow-me/rainbowkit": "^2.1.4",
+    "@tanstack/react-query": "^5.51.21",
     "@wagmi/connectors": "4.1.18",
     "@worldcoin/idkit": "^1.2.2",
+    "encoding": "^0.1.13",
     "ethers": "^5.7.2",
-    "next": "^13.4.2",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
+    "next": "^14.2.5",
+    "react": "18.3.1",
+    "react-dom": "18.3.1",
     "viem": "2.8.18",
     "wagmi": "2.5.12"
   },
   "devDependencies": {
-    "@types/node": "^20.2.1",
-    "@types/react": "^18.2.6",
-    "@types/react-dom": "^18.2.4",
-    "autoprefixer": "^10.4.14",
-    "eslint": "^8.40.0",
-    "eslint-config-next": "^13.4.2",
-    "postcss": "^8.4.23",
-    "tailwindcss": "^3.3.2",
-    "typescript": "^5.0.4"
+    "@types/node": "^20.14.14",
+    "@types/react": "^18.3.3",
+    "@types/react-dom": "^18.3.0",
+    "autoprefixer": "^10.4.20",
+    "eslint": "^8.57.0",
+    "eslint-config-next": "^13.5.6",
+    "pino-pretty": "^11.2.2",
+    "postcss": "^8.4.41",
+    "tailwindcss": "^3.4.7",
+    "typescript": "^5.5.4"
   },
   "browser": {
     "net": false
